requirements.txt
matplotlib==3.3.4
tqdm==4.61.2
imageio==2.9.0
numpy==1.20.3
opencv_python==4.5.3.56
Pillow==8.3.1
skimage
h5py